+++ tests/java/src/fr/ill/ics/cameo/test/Linked.java
@@ -16,6 +16,7 @@
 
 package fr.ill.ics.cameo.test;
 
+import fr.ill.ics.cameo.api.base.Option;
 import fr.ill.ics.cameo.api.base.Server;
 import fr.ill.ics.cameo.api.base.This;
 
@@ -35,7 +36,7 @@ public class Linked {
 			endpoint = "tcp://localhost:10000";
 		}
 		
-		Server server = Server.create(endpoint, useProxy);
+		Server server = Server.create(endpoint, (useProxy ? Option.USE_PROXY : 0));
 		server.init();
 		
 		// Start the application.