+++ tests/java/src/fr/ill/ics/cameo/test/TestAppExceptions.java
@@ -18,6 +18,7 @@ package fr.ill.ics.cameo.test;
 
 import fr.ill.ics.cameo.api.base.App;
 import fr.ill.ics.cameo.api.base.InitException;
+import fr.ill.ics.cameo.api.base.Option;
 import fr.ill.ics.cameo.api.base.Server;
 import fr.ill.ics.cameo.api.base.StartException;
 import fr.ill.ics.cameo.api.base.This;
@@ -38,7 +39,7 @@ public class TestAppExceptions {
 			endpoint = "tcp://localhost:10000";
 		}
 		
-		Server server = Server.create(endpoint, useProxy);
+		Server server = Server.create(endpoint, (useProxy ? Option.USE_PROXY : 0));
 		server.init();
 		
 		try {