+++ tests/java/src/fr/ill/ics/cameo/test/TestBasicRequester.java
@@ -17,6 +17,7 @@
 package fr.ill.ics.cameo.test;
 
 import fr.ill.ics.cameo.api.base.App;
+import fr.ill.ics.cameo.api.base.Option;
 import fr.ill.ics.cameo.api.base.Server;
 import fr.ill.ics.cameo.api.base.This;
 
@@ -52,7 +53,7 @@ public class TestBasicRequester {
 			endpoint = "tcp://localhost:10000";
 		}
 		
-		Server server = Server.create(endpoint, useProxy);
+		Server server = Server.create(endpoint, (useProxy ? Option.USE_PROXY : 0));
 		server.init();
 		
 		try {