+++ tests/java/src/fr/ill/ics/cameo/test/TestError.java
@@ -19,6 +19,7 @@ package fr.ill.ics.cameo.test;
 import java.util.Date;
 
 import fr.ill.ics.cameo.api.base.App;
+import fr.ill.ics.cameo.api.base.Option;
 import fr.ill.ics.cameo.api.base.Server;
 import fr.ill.ics.cameo.api.base.State;
 import fr.ill.ics.cameo.api.base.This;
@@ -45,7 +46,7 @@ public class TestError {
 			endpoint = "tcp://localhost:10000";
 		}
 		
-		Server server = Server.create(endpoint, useProxy);
+		Server server = Server.create(endpoint, (useProxy ? Option.USE_PROXY : 0));
 		server.init();		
 		
 		Date d = new Date();