+++ tests/java/src/fr/ill/ics/cameo/test/TestStream.java
@@ -38,7 +38,7 @@ public class TestStream {
 			endpoint = "tcp://localhost:10000";
 		}
 		
-		Server server = Server.create(endpoint, useProxy);
+		Server server = Server.create(endpoint, (useProxy ? Option.USE_PROXY : 0));
 		server.init();
 		
 		// Start the application.