+++ tests/java/src/fr/ill/ics/cameo/test/TestSubscriber.java
@@ -20,6 +20,7 @@ import java.util.List;
 
 import fr.ill.ics.cameo.api.base.App;
 import fr.ill.ics.cameo.api.base.InitException;
+import fr.ill.ics.cameo.api.base.Option;
 import fr.ill.ics.cameo.api.base.Server;
 import fr.ill.ics.cameo.api.base.This;
 
@@ -65,7 +66,7 @@ public class TestSubscriber {
 			endpoint = "tcp://localhost:10000";
 		}
 		
-		Server server = Server.create(endpoint, useProxy);
+		Server server = Server.create(endpoint, (useProxy ? Option.USE_PROXY : 0));
 		server.init();
 		
 		try {