+++ tests/java/src/fr/ill/ics/cameo/test/TestVeryFast.java
@@ -45,7 +45,7 @@ public class TestVeryFast {
 			endpoint = "tcp://localhost:10000";
 		}
 		
-		Server server = Server.create(endpoint, useProxy);
+		Server server = Server.create(endpoint, (useProxy ? Option.USE_PROXY : 0));
 		server.init();
 		
 		Date d = new Date();